pyproject.toml
[project]
name = "yt-dl-subtitle"
version = "0.3.0"
description = "YouTube subtitle downloader CLI tool"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "pytubefix",
    "python-dotenv",
    "asyncio"
]

[build-system]
requires = ["pyinstaller", "setuptools"]
build-backend = "setuptools.build_meta"

[project.scripts]
yt-dl-subtitle = "cli:main"
